# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Attendance System API
  version: 0.1.0
  description: |
    Multi-tenant API scoped by `company_id`.
    When authentication is enabled, tenant is derived from credentials (for example API key context).
    Request-scoped `company_id` parameters are primarily for development/testing and non-auth deployments.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/agent/bootstrap:
    post:
      tags: [agent-bridge]
      operationId: bootstrapAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentBootstrapRequest"
      responses:
        "201":
          description: Agent registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentBootstrapResponse"
        "400":
          description: Invalid token or payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/agent/heartbeat:
    post:
      tags: [agent-bridge]
      operationId: postAgentHeartbeat
      security:
        - AgentBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentHeartbeatRequest"
      responses:
        "200":
          description: Heartbeat accepted
          content:
            application/json:
              schema:
                type: object
        "503":
          description: Runtime schema is not ready for capability/reported-state writes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/agent/commands/next:
    post:
      tags: [agent-bridge]
      operationId: fetchAgentNextCommand
      security:
        - AgentBearerAuth: []
      responses:
        "200":
          description: Next command
          content:
            application/json:
              schema:
                type: object
  /api/agent/discovery-reports:
    post:
      tags: [agent-bridge]
      operationId: submitAgentDiscoveryReport
      security:
        - AgentBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscoveryReportRequest"
      responses:
        "201":
          description: Report accepted
          content:
            application/json:
              schema:
                type: object
  /api/agent/device-validations/result:
    post:
      tags: [agent-bridge]
      operationId: submitAgentDeviceValidationResult
      security:
        - AgentBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentDeviceValidationResultRequest"
      responses:
        "201":
          description: Validation result accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDeviceValidationResultResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Validation command not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Lifecycle conflict (command not pending, uid mismatch, or superseded uid without canonical)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/agent/device-events/batch:
    post:
      tags: [agent-bridge]
      operationId: submitAgentDeviceEventsBatch
      security:
        - AgentBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentDeviceEventsBatchRequest"
      responses:
        "201":
          description: Batch accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDeviceEventsBatchResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Device not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          description: Runtime schema is not ready for capability/reported-state writes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    AgentBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
...
  schemas:
...
    AgentDeviceValidationResultRequest:
      type: object
      required:
        - command_id
        - device_uid
        - completed_at
        - result
      properties:
        command_id:
          type: string
          format: uuid
        run_id:
          type: string
        device_uid:
          type: string
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
        result:
          type: object
          required:
            - success
          properties:
            success:
              type: boolean
            reason_code:
              type: string
              nullable: true
            evidence:
              type: object
    AgentDeviceValidationResultResponse:
      type: object
      properties:
        status:
          type: string
          enum: [ok]
        command_id:
          type: string
          format: uuid
        command_status:
          type: string
          enum: [acknowledged]
        requested_device_uid:
          type: string
        canonical_device_uid:
          type: string
        canonical_resolved_by:
          type: string
          enum: [direct, alias, superseded_redirect, unresolved]
        canonical_alias_kind:
          type: string
          nullable: true
        canonical_alias_status:
          type: string
          nullable: true
        result:
          type: object
          properties:
            success:
              type: boolean
            reason_code:
              type: string
              nullable: true
        lifecycle_event_written:
          type: boolean
        configuration_status:
          type: string
          enum: [complete, incomplete]
        missing_required_fields:
          type: array
          items:
            type: string
        ready_for_validation:
          type: boolean
        validation_status:
          type: string
          enum: [never_run, queued, in_progress, failed, succeeded]
        validation_reason_code:
          type: string
          nullable: true
        onboarding_state:
          type: string
          enum:
            - candidate_unconfigured
            - candidate_configured
            - validation_failed
            - validation_succeeded
            - managed_unbound
            - managed_pull_ready
...
    AgentHeartbeatRequest:
      type: object
      properties:
        uptime_sec:
          type: integer
          minimum: 0
        memory:
          type: object
        loadavg:
          type: array
          items:
            type: number
        hostname:
          type: string
        platform:
          type: string
        agent_version:
          type: string
        rollout_channel:
          type: string
          enum: [stable, beta]
    AgentBootstrapRequest:
      type: object
      required:
        - provisioning_token
        - agent_name
      properties:
        provisioning_token:
          type: string
        agent_name:
          type: string
        metadata:
          type: object
    AgentBootstrapResponse:
      type: object
      properties:
        agent_id:
          type: string
          format: uuid
        company_id:
          type: string
        agent_name:
          type: string
        credential_version:
          type: integer
        active_runtime_identity_id:
          type: string
          format: uuid
          nullable: true
        identity_status:
          type: string
          enum: [bootstrap_only, active, revoked, replaced]
          nullable: true
        runtime_identity_issued_at:
          type: string
          format: date-time
          nullable: true
        auth_token:
          type: string
        registered_at:
          type: string
    DiscoveryReportRequest:
      type: object
      required:
        - adapter_id
        - reported_at
        - discovered_devices
      properties:
        command_id:
          type: string
          format: uuid
        adapter_id:
          type: string
        subnet_targets:
          type: array
          items:
            type: string
        reported_at:
          type: string
        discovered_devices:
          type: array
          items:
            type: object
            required:
              - ip
              - vendor
              - observed_at
            properties:
              ip:
                type: string
              mac:
                type: string
              vendor:
                type: string
              model:
                type: string
              serial_number:
                type: string
              device_uid:
                type: string
              discovery_method:
                type: string
              confidence:
                type: number
              confidence_class:
                type: string
                enum: [high, medium, low]
              confirmation_state:
                type: string
                enum: [confirmed, zk_service_reachable, host_reachable]
              outcome_class:
                type: string
                enum: [confirmed, probable, heuristic]
              failure_reason:
                type: string
                enum: [timeout, auth_required, auth_failed, protocol_error, malformed_response, socket_error, network_unreachable]
                nullable: true
              identity_source:
                type: string
              protocol:
                type: object
              observed_at:
                type: string
              raw:
                type: object
        summary:
          type: object
    AgentDeviceEventsBatchRequest:
      type: object
      required:
        - device_uid
        - events
      properties:
        command_id:
          type: string
          format: uuid
        delivery_id:
          type: string
          description: Stable delivery identifier for idempotent retry/reconnect submission.
        delivery_attempt:
          type: integer
          minimum: 1
        run_id:
          type: string
        device_uid:
          type: string
        vendor:
          type: string
        ingest_method:
          type: string
          enum: [agent_pull, agent_realtime, agent_push]
        device_timezone:
          type: string
        pull_started_at:
          type: string
        pull_completed_at:
          type: string
          format: date-time
        cursor:
          type: object
          properties:
            requested_since_utc:
              type: string
              format: date-time
            latest_event_time_utc:
              type: string
              format: date-time
            has_more:
              type: boolean
        summary:
          type: object
        events:
          type: array
          items:
            type: object
            required:
              - device_person_id
              - event_time_local
              - event_time_utc
              - direction
            properties:
              device_person_id:
                type: string
              event_time_local:
                type: string
              event_time_utc:
                type: string
                format: date-time
              device_timezone:
                type: string
              direction:
                type: string
                enum: [IN, OUT]
              event_type:
                type: string
              verify_state:
                type: string
              verify_method:
                type: string
              dedup_key:
                type: string
              raw:
                type: object
    AgentDeviceEventsBatchResponse:
      type: object
      properties:
        status:
          type: string
        batch_id:
          type: string
          format: uuid
        delivery_id:
          type: string
        replayed_delivery:
          type: boolean
        inserted_count:
          type: integer
        deduped_count:
          type: integer
        rejected_count:
          type: integer
        latest_event_time_utc:
          type: string
        rejection_samples:
          type: array
          items:
            type: object
...
    ErrorEnvelope:
      type: object
      required:
        - error
        - code
        - message
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
        details: {}